requirements.txt
watchdog
requests
notion-client
python-dotenv